requirements.txt
# openstack_inventory
python-novaclient>=7.0

# openstack_upload_metadata.py
ansible >= 2.0
